Nvidia Isaac sim wheelchair simulation
The wheelchair's autonomous navigation using waypoint is performed using the Isaac sim. The Isaac Sim is a robot simulation tool provided by the Nvidia Omniverse. The Isaac sim has a ROS bridge allowing it to send rostopics from the simulation software to the ROS framework.

The folders "wheelchair_description" and "wheelchair_navigation" can be git cloned onto the src folder of the ros_workspace. This ros_workspace is the default workspace for the Nvidia Omniverse.